# openapi.yaml
openapi: 3.0.3
info:
  title: Enhanced Orders API
  description: |
    A comprehensive orders management API built with AWS Lambda and serverless.tf modules.

    This API demonstrates advanced patterns from the aws-lambda-handler-cookbook including:
    - Three-layer architecture (handlers, logic, dal)
    - Lambda Powertools with idempotency
    - AWS AppConfig for feature flags
    - Comprehensive error handling and validation
    - Full observability (logs, metrics, tracing)

    ## Features
    - Order lifecycle management (create, read, update, cancel)
    - Real-time health monitoring
    - Business metrics and statistics
    - Idempotent operations
    - Comprehensive error responses
    - CORS support

    ## Authentication
    Currently no authentication is required, but the API is designed to easily integrate
    with AWS Cognito, API Keys, or custom authorizers.

  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.example.com/api/v1
    description: Production server
  - url: https://staging-api.example.com/api/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Local development server
...
paths:
...
  /orders:
    get:
      tags:
        - orders
      summary: List orders
      description: |
        Retrieve a paginated list of orders with optional filtering.
        Supports filtering by customer email and order status.
      operationId: listOrders
      parameters:
        - name: customer_email
          in: query
          description: Filter orders by customer email
          required: false
          schema:
            type: string
            format: email
          example: "john.doe@example.com"
        - name: status
          in: query
          description: Filter orders by status
          required: false
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - name: limit
          in: query
          description: Maximum number of orders to return (1-100)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: last_evaluated_key
          in: query
          description: Pagination token for retrieving next page
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of orders retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrdersList'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service temporarily unavailable (maintenance mode)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - orders
      summary: Create a new order
      description: |
        Create a new order with automatic validation and business rule enforcement.
        This operation is idempotent based on request body content.
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            examples:
              basic_order:
                summary: Basic order
                value:
                  customer_name: "John Doe"
                  customer_email: "john.doe@example.com"
                  order_item_count: 5
              order_with_notes:
                summary: Order with notes
                value:
                  customer_name: "Jane Smith"
                  customer_email: "jane.smith@example.com"
                  order_item_count: 3
                  notes: "Please deliver after 5 PM"
      responses:
        '201':
          description: Order created successfully
          headers:
            Location:
              description: URL of the created order
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateOrderResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Request already in progress (idempotency collision)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Business logic validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service temporarily unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{order_id}:
    parameters:
      - name: order_id
        in: path
        required: true
        description: Unique identifier for the order
        schema:
          type: string
          format: uuid
        example: "550e8400-e29b-41d4-a716-446655440000"

    get:
      tags:
        - orders
      summary: Get order by ID
      description: Retrieve detailed information about a specific order
      operationId: getOrderById
      responses:
        '200':
          description: Order retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetOrderResponse'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Service temporarily unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - orders
      summary: Update an existing order
      description: |
        Update order details. Only orders in certain statuses can be updated.
        This operation is idempotent based on request body content.
      operationId: updateOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderRequest'
            examples:
              update_quantity:
                summary: Update item count
                value:
                  order_item_count: 8
              update_notes:
                summary: Update notes
                value:
                  notes: "Updated delivery instructions: Leave at front door"
              update_both:
                summary: Update quantity and notes
                value:
                  order_item_count: 10
                  notes: "Rush order - expedite processing"
      responses:
        '200':
          description: Order updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateOrderResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Request already in progress (idempotency collision)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Order cannot be updated in current status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - orders
      summary: Cancel an order
      description: |
        Cancel an order if it's in a cancellable status.
        Cancelled orders cannot be restored.
      operationId: cancelOrder
      parameters:
        - name: reason
          in: query
          description: Reason for cancelling the order
          required: false
          schema:
            type: string
            maxLength: 200
            default: "Customer requested cancellation"
          example: "Customer changed mind"
      responses:
        '200':
          description: Order cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateOrderResponse'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Order cannot be cancelled in current status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/statistics:
    get:
      tags:
        - statistics
      summary: Get order statistics
      description: |
        Retrieve comprehensive business metrics and statistics about orders.
        Useful for monitoring, reporting, and business intelligence.
      operationId: getOrderStatistics
      responses:
        '200':
          description: Statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderStatistics'
        '503':
          description: Service temporarily unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    OrderStatus:
      type: string
      enum:
        - pending
        - confirmed
        - processing
        - shipped
        - delivered
        - cancelled
        - refunded
      description: Current status of the order
      example: pending
...
    CreateOrderRequest:
      type: object
      required:
        - customer_name
        - customer_email
        - order_item_count
      properties:
        customer_name:
          type: string
          minLength: 1
          maxLength: 50
          pattern: "^[a-zA-Z\\s]+$"
          description: Customer name for the order
          example: "John Doe"
        customer_email:
          type: string
          format: email
          description: Customer email address
          example: "john.doe@example.com"
        order_item_count:
          type: integer
          minimum: 1
          maximum: 100
          description: Number of items in the order
          example: 5
        notes:
          type: string
          maxLength: 500
          description: Optional notes for the order
          example: "Please deliver after 5 PM"
    UpdateOrderRequest:
      type: object
      properties:
        order_item_count:
          type: integer
          minimum: 1
          maximum: 100
          description: Updated number of items in the order
          example: 8
        notes:
          type: string
          maxLength: 500
          description: Updated notes for the order
          example: "Updated delivery instructions"
    CreateOrderResponse:
      type: object
      required:
        - order_id
        - status
        - total_amount
        - created_at
        - message
      properties:
        order_id:
          type: string
          format: uuid
          description: Unique identifier for the created order
          example: "550e8400-e29b-41d4-a716-446655440000"
        status:
          $ref: '#/components/schemas/OrderStatus'
        total_amount:
          type: number
          format: float
          minimum: 0
          description: Total order amount in USD
          example: 250.0
        created_at:
          type: string
          format: date-time
          description: ISO timestamp when the order was created
          example: "2024-01-15T10:30:00Z"
        message:
          type: string
          description: Success message
          example: "Order created successfully"
    GetOrderResponse:
      type: object
      required:
        - order
      properties:
        order:
          $ref: '#/components/schemas/Order'
    UpdateOrderResponse:
      type: object
      required:
        - order
        - message
      properties:
        order:
          $ref: '#/components/schemas/Order'
        message:
          type: string
          description: Success message
          example: "Order updated successfully"
    OrdersList:
      type: object
      required:
        - orders
        - count
        - scanned_count
      properties:
        orders:
          type: array
          items:
            $ref: '#/components/schemas/Order'
          description: List of orders
        count:
          type: integer
          minimum: 0
          description: Number of orders returned in this response
          example: 25
        scanned_count:
          type: integer
          minimum: 0
          description: Number of orders scanned to generate this response
          example: 25
        last_evaluated_key:
          type: object
          description: Pagination token for retrieving the next page
          additionalProperties: true
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - error_id
            - timestamp
          properties:
            code:
              type: string
              description: Error code for programmatic handling
              example: "VALIDATION_ERROR"
            message:
              type: string
              description: Human-readable error message
              example: "Invalid input provided. Please check your request and try again."
            error_id:
              type: string
              description: Unique error identifier for tracking
              example: "error-12345-67890"
            timestamp:
              type: string
              format: date-time
              description: ISO timestamp when error occurred
              example: "2024-01-15T10:30:00Z"
            details:
              type: object
              description: Additional error context
              properties:
                operation:
                  type: string
                  example: "create_order"
                resource_id:
                  type: string
                  example: "550e8400-e29b-41d4-a716-446655440000"
            field_errors:
              type: array
              description: Field-specific validation errors
              items:
                type: object
                properties:
                  field:
                    type: string
                    description: Field name with error
                    example: "customer_email"
                  message:
                    type: string
                    description: Field error message
                    example: "Invalid email format"
        retry_after:
          type: integer
          description: Seconds to wait before retrying (for rate limiting)
          example: 60
...